chore: Update Hero and Navbar components
The Hero component in Hero.tsx is updated to improve the layout and styling of the heading. The text size is adjusted to be responsive, and the line break is added between "Altın Saray" and "Wedding Hall".

The Navbar component in Navbar.tsx is updated to include the "hidden" class on small screens, ensuring it is not visible. This change is made to improve the responsiveness of the navigation bar.

These changes are made to enhance the visual appearance and user experience of the website.